requirements.txt
sphinxcontrib-napoleon==0.2.9
plyer=1.4.2
numpy=1.17.3
pyproj=2.4.0
exifread=2.1.2
pyexif=1.1.3
unidecode=1.1.1
pillow=6.2.0
